qe: absorb SqlInfo in Context
The two types have very similar use and are used in the same places. We
can simplify by merging them.
