+++ query-engine/connectors/sql-query-connector/src/database/transaction.rs
@@ -1,5 +1,5 @@
 use super::catch;
-use crate::{database::operations::*, sql_info::SqlInfo, Context, SqlError};
+use crate::{database::operations::*, Context, SqlError};
 use async_trait::async_trait;
 use connector::{ConnectionLike, RelAggregationSelection};
 use connector_interface::{
@@ -102,7 +102,6 @@ impl<'tx> ReadOperations for SqlConnectorTransaction<'tx> {
                 query_arguments,
                 &selected_fields.into(),
                 aggr_selections,
-                SqlInfo::from(&self.connection_info),
                 &ctx,
             )
             .await
@@ -164,15 +163,7 @@ impl<'tx> WriteOperations for SqlConnectorTransaction<'tx> {
     ) -> connector::Result<usize> {
         catch(self.connection_info.clone(), async move {
             let ctx = Context::new(&self.connection_info, trace_id.as_deref());
-            write::create_records(
-                &self.inner,
-                SqlInfo::from(&self.connection_info),
-                model,
-                args,
-                skip_duplicates,
-                &ctx,
-            )
-            .await
+            write::create_records(&self.inner, model, args, skip_duplicates, &ctx).await
         })
         .await
     }
@@ -273,7 +264,7 @@ impl<'tx> WriteOperations for SqlConnectorTransaction<'tx> {
         _query_type: Option<String>,
     ) -> connector::Result<serde_json::Value> {
         catch(self.connection_info.clone(), async move {
-            write::query_raw(&self.inner, SqlInfo::from(&self.connection_info), self.features, inputs).await
+            write::query_raw(&self.inner, inputs).await
         })
         .await
     }
